requirements.txt
fastapi
streamlit
torch
torchvision
pillow
uvicorn
requests
